Extract KuduColumnProperties from KuduTableProperties
